Add filename, Cleanup, Finish seperating JavaScript from the Scripting Language Compiler
