+++ src/main/java/io/github/ngspace/hudder/compilers/EmptyCompiler.java
@@ -6,7 +6,7 @@ import io.github.ngspace.hudder.config.ConfigInfo;
 
 public class EmptyCompiler extends ATextCompiler {
 	@Override
-	public CompileResult compile(ConfigInfo info, String text) throws CompileException {return CompileResult.of(text);}
+	public CompileResult compile(ConfigInfo info, String text, String filename) throws CompileException {return CompileResult.of(text);}
 
 	@Override public Object getVariable(String key) throws CompileException {return null;}
 }
