+++ src/main/java/io/github/ngspace/hudder/methods/methods/LoadMethod.java
@@ -22,7 +22,7 @@ public class LoadMethod implements IMethod {
 			boolean AddText = (args.length<2 || args[1].asBoolean()) || type.equals("add");
 			ATextCompiler ecompiler=(args.length>2?Compilers.getCompilerFromName(args[2].asString()):comp);
 			for (var i : Hudder.precomplistners) i.accept(ecompiler);
-			meta.combineWithResult(ecompiler.compile(ci, HudFileUtils.getFile(file)), AddText);
+			meta.combineWithResult(ecompiler.compile(ci, HudFileUtils.getFile(file), file), AddText);
 			for (var i : Hudder.postcomplistners) i.accept(ecompiler);
 		} catch (ReflectiveOperationException | IOException e) {
 			throw new CompileException(e.getLocalizedMessage());
