+++ src/main/java/io/github/ngspace/hudder/util/testing/HudderUnitTest.java
@@ -22,7 +22,7 @@ public class HudderUnitTest {
 	public HudderUnitTestResult test(ConfigInfo info) {
 		try {
 			if (compiler instanceof AV2Compiler v2comp) v2comp.runtimes = new HashMap<String, V2Runtime>();
-			String text = compiler.compile(info, texttocompile).TopLeftText;
+			String text = compiler.compile(info, texttocompile, "Unit Tests").TopLeftText;
 			boolean res = expectation.equals(text);
 			return new HudderUnitTestResult(res, expectation, text.replaceAll("(^ )|( $)", "~"));
 		} catch (Exception e) {
